+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- Version Index -->
+* [0.13.0](#0.13.0)
 * [0.10.6](#0.10.6)
 * [0.10.4](#0.10.4)
 * [0.10.3](#0.10.3)
@@ -14,7 +15,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [0.9.2](#0.9.2)
 
 <!-- Changelog Description -->
-## 0.10.6
+##0.13.0
+
+### Added
+* Flag when element is updated.
+* Element on update function.
+* Get element function.
+* Option to await variable on change.
+
+### Changed
+* Denv set function now sets updated flag.
+* denv_shmem_destroy no longer needs table, only shmem bind path.
+
+### Fixed
+* Removing table no longer loads the table (useful for removing broken tables).
+
+##0.10.6
 
 ### Added
 * Flag to force yes to operations that prompts the user.
@@ -25,23 +41,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Bug when loading to a different path.
 
-## 0.10.4
+##0.10.4
 
 ### Fixed
 * Bug when trying to get a variable that doesn't exist.
 
-## 0.10.3
+##0.10.3
 
 ### Fixed
 * Denv not creating the path for it's default shared environment.
 
-## 0.10.2
+##0.10.2
 
 ### Fixed
 * Bug when removing denv instances both from denv default bind path.
 * Bug when removing denv from another bind path.
 
-## 0.10.0
+##0.10.0
 
 ### Added
 * Contributors file [CONTRIBUTORS.md](CONTRIBUTORS.md).
@@ -52,7 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Commands are now full words, options are set as flags.
 * License is now GPL-3.
 
-## 0.9.2
+##0.9.2
 
 ### Added
 * This changelog file.